move web worker init from editor.js to index.js
