+++ src/input/editor.js
@@ -1,4 +1,4 @@
-import { runGameCommands, toggleGrid } from "../index.js";
+import { initializeWorker } from "../index.js";
 
 //Lint cheese below
 /* global ace */
@@ -24,13 +24,7 @@ function initializeEditor() {
         autoScrollEditorIntoView: true,
         copyWithEmptySelection: true,
     });
-    //Create "run code" button
-    let buttonID = "editor-button";
-    //CreateButton(buttonID, "Suorita");
-    addEventToButton(buttonID);
-
-    // Toggle grid nappi, voi muuttaa missä annetaan event
-    document.getElementById("grid-toggle-button").addEventListener("click", toggleGrid, false);
+    addEventToButton("editor-button");
 }
 
 
@@ -40,26 +34,5 @@ function addEventToButton(id) {
 }
 
 function onClickCodeButton() {
-    const worker = new Worker('src/input/pyodide.js');
-
-    worker.onmessage = function (event) {
-        if (event.data.type === 'input') {
-            const sharedArray = new Uint16Array(event.data.sab, 4);
-            const syncArray = new Int32Array(event.data.sab, 0, 1);
-
-            const word = prompt(event.data.message);
-
-            for (let i = 0; i < word.length; i++) {
-                sharedArray[i] = word.charCodeAt(i);
-            }
-            sharedArray[word.length] = 0;
-
-            Atomics.store(syncArray, 0, 1);
-            Atomics.notify(syncArray, 0, 1);
-        } else if (event.data.type === 'run') {
-            runGameCommands(event.data.data);
-        }
-    }
-
-    worker.postMessage({ type: 'start', data: editor.getValue() });
+    initializeWorker(editor);
 }
